i added... stuff
i started off with this one expression generator because i thought math training was literally well... training your basic math skills, so that file is expressionGenerator.java and it turns out the questions does not need to be randomly generated so its just kinda there, but since i can put it in a forloop in like the tryout file js to generate as many questions as possible, now i have 50 random questions for basics.txt and a file that contributes basically nothing to the main file

i also changed the format of answer? because when betaing the game and letting players play, more students typed +7 instead of plus seven or positive seven so i made it more practical.